Structure files to make room for has_many
