+++ lib/superstore/associations/builder/belongs_to.rb
@@ -1,0 +1,4 @@
+module Superstore::Associations::Builder
+  class BelongsTo < Association
+  end
+end
