+++ lib/superstore/associations/builder/has_many.rb
@@ -1,0 +1,4 @@
+module Superstore::Associations::Builder
+  class HasMany < Association
+  end
+end
